backend/requirements.txt
# FastAPI核心
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# 富途SDK
futu-api==6.6.0

# 数据库
sqlalchemy==2.0.25
aiosqlite==0.19.0

# 数据验证
pydantic==2.5.3
pydantic-settings==2.1.0

# WebSocket
websockets==12.0

# 工具
python-dotenv==1.0.0
httpx==0.26.0

# 测试
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
